# contracts/openapi/mesh-benchmark-api.yaml
openapi: 3.1.0
info:
  title: Mesh Performance Lab Data Plane API
  version: 1.0.0
  description: Measurement-path contracts. Load is generated by k6, not the web console.
  license: { name: UNLICENSED, identifier: LicenseRef-Proprietary }
servers:
  - url: /api/v1
security: []
paths:
  /system/ping:
    get:
      operationId: pingSystem
      summary: Smoke-test gateway and orchestrator connectivity
      responses:
        '200':
          description: Connectivity result
          content: { application/json: { schema: { $ref: '#/components/schemas/PingResult' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /workloads/chain:
    post:
      operationId: executeChain
      summary: Execute a synchronous hop chain
      parameters: [{ $ref: '#/components/parameters/RunId' }, { $ref: '#/components/parameters/CorrelationId' }, { $ref: '#/components/parameters/Deadline' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ChainRequest' } } }
      responses:
        '200':
          description: Chain completed
          content: { application/json: { schema: { $ref: '#/components/schemas/ChainResult' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '504': { $ref: '#/components/responses/DeadlineExceeded' }
  /workloads/fanout:
    post:
      operationId: executeFanout
      summary: Execute sequential or parallel fan-out
      parameters: [{ $ref: '#/components/parameters/RunId' }, { $ref: '#/components/parameters/CorrelationId' }, { $ref: '#/components/parameters/Deadline' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/FanoutRequest' } } }
      responses:
        '200':
          description: Complete or partial fan-out result
          content: { application/json: { schema: { $ref: '#/components/schemas/FanoutResult' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /workloads/target:
    post:
      operationId: executeTargetWork
      summary: Execute bounded delay, error, CPU, memory and I/O work
      parameters: [{ $ref: '#/components/parameters/RunId' }, { $ref: '#/components/parameters/CorrelationId' }, { $ref: '#/components/parameters/Deadline' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/TargetWorkRequest' } } }
      responses:
        '200':
          description: Work completed
          content: { application/json: { schema: { $ref: '#/components/schemas/TargetWorkResult' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/InjectedFailure' }
        '504': { $ref: '#/components/responses/DeadlineExceeded' }
  /workloads/payload:
    post:
      operationId: exchangePayload
      summary: Generate or echo a deterministic payload
      parameters: [{ $ref: '#/components/parameters/RunId' }, { $ref: '#/components/parameters/CorrelationId' }, { $ref: '#/components/parameters/Deadline' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PayloadRequest' } } }
      responses:
        '200':
          description: Payload metadata and optional body
          content: { application/json: { schema: { $ref: '#/components/schemas/PayloadResult' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '504': { $ref: '#/components/responses/DeadlineExceeded' }
  /workloads/async/tasks:
    post:
      operationId: publishTasks
      summary: Publish deterministic benchmark tasks to Kafka
      parameters: [{ $ref: '#/components/parameters/RunId' }, { $ref: '#/components/parameters/CorrelationId' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/PublishTasksRequest' } } }
      responses:
        '202':
          description: Batch accepted for publication
          content: { application/json: { schema: { $ref: '#/components/schemas/PublishTasksResult' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '503': { $ref: '#/components/responses/Unavailable' }
components:
  parameters:
    RunId:
      name: X-Experiment-Run-Id
      in: header
      required: true
      schema: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._-]{0,99}$' }
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 100 }
    Deadline:
      name: X-Request-Deadline-Epoch-Ms
      in: header
      required: false
      description: Absolute UTC epoch millisecond deadline; services propagate the remaining budget.
      schema: { type: integer, format: int64, minimum: 0 }
  responses:
    BadRequest:
      description: Invalid bounded workload configuration
      content: { application/json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Unavailable:
      description: Required downstream is unavailable
      content: { application/json: { schema: { $ref: '#/components/schemas/Problem' } } }
    DeadlineExceeded:
      description: Remaining request budget was exhausted
      content: { application/json: { schema: { $ref: '#/components/schemas/Problem' } } }
    InjectedFailure:
      description: Deterministic configured failure
      content: { application/json: { schema: { $ref: '#/components/schemas/Problem' } } }
  schemas:
    PingResult:
      type: object
      additionalProperties: false
      required: [service, status, traceId, downstream, timestamp]
      properties:
        service: { type: string }
        status: { const: UP }
        traceId: { type: string, pattern: '^[a-f0-9]{32}$' }
        downstream: { type: object, additionalProperties: true }
        timestamp: { type: string, format: date-time }
    WorkSpec:
      type: object
      additionalProperties: false
      properties:
        delayMs: { type: integer, minimum: 0, maximum: 60000, default: 0 }
        delayDistribution: { enum: [FIXED, NORMAL, EXPONENTIAL], default: FIXED }
        errorRate: { type: number, minimum: 0, maximum: 1, default: 0 }
        cpuMillis: { type: integer, minimum: 0, maximum: 10000, default: 0 }
        memoryBytes: { type: integer, minimum: 0, maximum: 67108864, default: 0 }
        blockingIoMs: { type: integer, minimum: 0, maximum: 10000, default: 0 }
        seed: { type: integer, format: int64, default: 0 }
    ChainRequest:
      type: object
      additionalProperties: false
      required: [hopCount, work]
      properties:
        hopCount: { type: integer, minimum: 0, maximum: 16 }
        payloadBytes: { type: integer, minimum: 0, maximum: 10485760, default: 0 }
        work: { $ref: '#/components/schemas/WorkSpec' }
    ChainResult:
      type: object
      additionalProperties: false
      required: [status, completedHops, elapsedMs, checksum]
      properties:
        status: { enum: [COMPLETED, FAILED, DEADLINE_EXCEEDED] }
        completedHops: { type: integer, minimum: 0 }
        elapsedMs: { type: number, minimum: 0 }
        checksum: { type: string }
    FanoutRequest:
      type: object
      additionalProperties: false
      required: [targetCount, mode, timeoutBudgetMs]
      properties:
        targetCount: { type: integer, minimum: 1, maximum: 64 }
        mode: { enum: [SEQUENTIAL, PARALLEL] }
        timeoutBudgetMs: { type: integer, minimum: 1, maximum: 60000 }
        allowPartial: { type: boolean, default: true }
        work: { $ref: '#/components/schemas/WorkSpec' }
    TargetOutcome:
      type: object
      additionalProperties: false
      required: [target, status, elapsedMs]
      properties:
        target: { type: string }
        status: { enum: [COMPLETED, FAILED, TIMED_OUT, CANCELLED] }
        elapsedMs: { type: number, minimum: 0 }
        errorCode: { type: [string, 'null'] }
    FanoutResult:
      type: object
      additionalProperties: false
      required: [status, completedTargets, failedTargets, outcomes]
      properties:
        status: { enum: [COMPLETED, PARTIAL, FAILED] }
        completedTargets: { type: integer, minimum: 0 }
        failedTargets: { type: integer, minimum: 0 }
        outcomes: { type: array, items: { $ref: '#/components/schemas/TargetOutcome' }, maxItems: 64 }
    TargetWorkRequest:
      type: object
      additionalProperties: false
      required: [work]
      properties:
        work: { $ref: '#/components/schemas/WorkSpec' }
        responseBytes: { type: integer, minimum: 0, maximum: 10485760, default: 0 }
    TargetWorkResult:
      type: object
      additionalProperties: false
      required: [instance, elapsedMs, checksum]
      properties:
        instance: { type: string }
        elapsedMs: { type: number, minimum: 0 }
        allocatedBytes: { type: integer, minimum: 0 }
        checksum: { type: string }
    PayloadRequest:
      type: object
      additionalProperties: false
      required: [sizeBytes, seed, mode]
      properties:
        sizeBytes: { type: integer, minimum: 0, maximum: 10485760 }
        seed: { type: integer, format: int64 }
        mode: { enum: [BUFFERED, STREAMING] }
        compression: { enum: [IDENTITY, GZIP], default: IDENTITY }
        includeBody: { type: boolean, default: true }
    PayloadResult:
      type: object
      additionalProperties: false
      required: [sizeBytes, checksum, mode, compression]
      properties:
        sizeBytes: { type: integer, minimum: 0 }
        checksum: { type: string }
        mode: { enum: [BUFFERED, STREAMING] }
        compression: { enum: [IDENTITY, GZIP] }
        bodyBase64: { type: [string, 'null'], contentEncoding: base64 }
    PublishTasksRequest:
      type: object
      description: taskCount multiplied by payloadBytes must not exceed the 64 MiB batch safety limit.
      additionalProperties: false
      required: [taskCount, processingMillis, payloadBytes, seed]
      properties:
        taskCount: { type: integer, minimum: 1, maximum: 10000 }
        processingMillis: { type: integer, minimum: 0, maximum: 60000 }
        payloadBytes: { type: integer, minimum: 0, maximum: 10485760 }
        seed: { type: integer, format: int64 }
    PublishTasksResult:
      type: object
      additionalProperties: false
      required: [batchId, accepted, rejected]
      properties:
        batchId: { type: string, format: uuid }
        accepted: { type: integer, minimum: 0 }
        rejected: { type: integer, minimum: 0 }
    Problem:
      type: object
      additionalProperties: false
      required: [code, message, correlationId, experimentRunId, retryable]
      properties:
        code: { type: string }
        message: { type: string }
        correlationId: { type: string }
        experimentRunId: { type: string }
        retryable: { type: boolean }
        target: { type: [string, 'null'] }